[REF] iot_drivers: rename static files using camel_case
We rename static files to camel_case to comply with Odoo
conventions.

We also reduced the timeout before the full screen loader
starts pinging the IoT Box. It was 10 seconds to avoid pinging
before the server stopped. We reduced it to 2 seconds, as it's
useless most of the time.
