+++ addons/iot_drivers/static/src/app/components/single_data.js
