+++ addons/iot_drivers/static/src/app/hooks/store_hook.js
